Rename peopleFormSelectors to personFormSelectors
This matches its Container name, although it still doesn't match
the actual component name of PersonInformationForm.
